refactor: move iterative training to analysis
